Rename include/Scrape to include/probe
