+++ include/notifier.php
@@ -3,9 +3,9 @@
 use Friendica\App;
 use Friendica\Core\Config;
 
-require_once('include/Scrape.php');
 require_once 'include/queue_fn.php';
 require_once 'include/html2plain.php';
+require_once 'include/probe.php';
 require_once 'include/diaspora.php';
 require_once 'include/ostatus.php';
 require_once 'include/salmon.php';
