+++ include/ostatus.php
@@ -5,7 +5,6 @@
 
 use Friendica\App;
 use Friendica\Core\Config;
-require_once("include/Scrape.php");
 use Friendica\Network\Probe;
 
 require_once 'include/Contact.php';
@@ -17,6 +16,7 @@ require_once 'mod/share.php';
 require_once 'include/enotify.php';
 require_once 'include/socgraph.php';
 require_once 'include/Photo.php';
+require_once 'include/probe.php';
 require_once 'include/follow.php';
 require_once 'include/api.php';
 require_once 'mod/proxy.php';
