+++ include/socgraph.php
@@ -9,10 +9,10 @@
 
 use Friendica\App;
 use Friendica\Core\Config;
-require_once("include/Scrape.php");
 use Friendica\Network\Probe;
 
 require_once 'include/datetime.php';
+require_once 'include/probe.php';
 require_once 'include/network.php';
 require_once 'include/html2bbcode.php';
 require_once 'include/Contact.php';
