+++ include/update_gcontact.php
@@ -5,7 +5,7 @@ use Friendica\Core\Config;
 function update_gcontact_run(&$argv, &$argc) {
 	global $a;
 
-	require_once 'include/Scrape.php';
+	require_once 'include/probe.php';
 	require_once 'include/socgraph.php';
 
 	logger('update_gcontact: start');
