+++ mod/contacts.php
@@ -3,10 +3,10 @@
 use Friendica\App;
 use Friendica\Network\Probe;
 
-require_once('include/Scrape.php');
 require_once 'include/Contact.php';
 require_once 'include/socgraph.php';
 require_once 'include/contact_selectors.php';
+require_once 'include/probe.php';
 require_once 'mod/proxy.php';
 require_once 'include/Photo.php';
