+++ mod/dfrn_request.php
@@ -13,10 +13,10 @@
  */
 
 use Friendica\App;
-
-require_once('include/Scrape.php');
 use Friendica\Network\Probe;
+
 require_once 'include/enotify.php';
+require_once 'include/probe.php';
 require_once 'include/group.php';
 
 function dfrn_request_init(App $a) {
