+++ mod/ostatus_subscribe.php
@@ -2,7 +2,7 @@
 
 use Friendica\App;
 
-require_once('include/Scrape.php');
+require_once 'include/probe.php';
 require_once 'include/follow.php';
 
 function ostatus_subscribe_content(App $a) {
