+++ mod/repair_ostatus.php
@@ -2,7 +2,7 @@
 
 use Friendica\App;
 
-require_once('include/Scrape.php');
+require_once 'include/probe.php';
 require_once 'include/follow.php';
 
 function repair_ostatus_content(App $a) {
